requirements.txt
flet>=0.84.0
genanki
python-dotenv
requests
pyinstaller
